requirements.txt
pathlib2>=2.3.5
opencc-python-reimplemented>=0.1.5
PyYAML>=5.2
fire>=0.2.1
